go-module.eclass: invoke "ego mod tidy" with nonfatal
We previously invoked "mod tidy" without the ego helper function in a
non-fatal way. This was switched to using the ego helper in
f9ee55e698f8 ("go-module.eclass: use ego helper function") but without
prefixing the call with nonfatal. Hence the semantic of the eclass
changes: a previously non-fatal invocation became fatal, causing e.g.,
bug #834695.

[sam: fixes semantic change which wasn't advertised/seemingly intentional,
so let's get this in now to avoid a significant number of ebuilds being broken.

We can introduce this change if desired at a later point.

Fixes: f9ee55e698f8 ("go-module.eclass: use ego helper function")
Closes: https://bugs.gentoo.org/834758
Closes: https://bugs.gentoo.org/834762
Closes: https://bugs.gentoo.org/834607
Closes: https://bugs.gentoo.org/834608
Closes: https://bugs.gentoo.org/834625
Closes: https://bugs.gentoo.org/834626
Closes: https://bugs.gentoo.org/834630
Closes: https://bugs.gentoo.org/834669
Closes: https://bugs.gentoo.org/834603
Closes: https://bugs.gentoo.org/834610
Closes: https://bugs.gentoo.org/834711
Closes: https://bugs.gentoo.org/834763
Signed-off-by: Florian Schmaus <flow@gentoo.org>
Closes: https://github.com/gentoo/gentoo/pull/24437
Signed-off-by: Sam James <sam@gentoo.org>