media-gfx/hydrus: Drop 469, add 476, update 9999
The documentation has been changed to use mkdocs instead of plain html,
therefore the ebuild had to be adapted for this. Additionally, the
license.txt file turned out to not be a duplicate, but instead used by
the program to show the license, it therefore has to be in the same
directory as the program and not in the doc directory.

Closes: https://bugs.gentoo.org/834921
Signed-off-by: Ekaterina Vaartis <vaartis@kotobank.ch>
Closes: https://github.com/gentoo/gentoo/pull/24471
Signed-off-by: Sam James <sam@gentoo.org>