media-gfx/inkscape: use internal 2geom explicitly
When inkscape is first time emerged to the system its 2geom library
detection mechanism selects bundled version and installs it as part of
the package. This detection mechanism can lead to a runtime problem when
inkscape is reemerged, because the 2geom library is located in the
system (from previous installation), therefore bundled version is not
selected nor appended to the package and emerge removes 2geom library
from the system because there is no replacement in a new package. Then,
inkscape fails to start because of missing library.

It is necessary to force use of the bundled 2geom library in order to
address the problem.

Signed-off-by: Petr Vaněk <arkamar@atlas.cz>