app-crypt/libsecret: Create an emulated TPM2 for tests
Libsecret runs tests against the TPM2 already in the machine and require
tpm2-abrmd. Hence the tests will fail if the user does not have tpm2-abrmd
installed or does not have a TPM2 in the machine. It shouldn't do this
- it should provision a virtual TPM2 spawn an emulator (swtpm) and the use
the swtpm TCTI with tpm2-abrmd.

However its not too difficult to setup the TPM2 simulator for the test,
which is what this patch does.

Bug: https://bugs.gentoo.org/834830
Signed-off-by: Christopher Byrne <salah.coronya@gmail.com>