media-video/ffmpeg: backport small GNU ranlib hardcoding fix
I've landed this fix in ffmpeg upstream and would like to backport
it to Gentoo so the --ranlib="$(tc-getRANLIB) option set in the
ebuilds has an effect, to be able to use the LLVM tool which is
currently ignored.

Because this is a small fix not impacting the majority of users
who don't care about mixed LLVM/GNU tools usage, I don't think
we nede to bump the ebuild revisons to cause rebuilds.

Closes: https://bugs.gentoo.org/835419
Signed-off-by: Adrian Ratiu <adrian.ratiu@collabora.com>