net-dns/bind: restore USE=jemalloc
We can't force jemalloc because bind-tools (which this now
blocks, and installs the same tools as, and we may end up
just using net-dns/bind for all of it) needs to be
usable in as many places as possible and jemalloc isn't
ported to all arches.

We can therefore restore ~sparc.

Bug: https://bugs.gentoo.org/832218
Signed-off-by: Sam James <sam@gentoo.org>