net-wireless/wpa_supplicant: fix wep regression
Introduce a use flag for wep.  Previous bump accidently left wep hard
disabled.  That's a very sane default, but we need a way to turn it back
on for anyone who needs it.
Fixes: 744939bdfc1a0a6296c709d4382d3676abdb2b66
Additionally add a use flag for TKIP which was introduced in 2.10.
While 2.10 leaves TKIP enabled by default, it's been deprecated since
2009 and support has been removed from many other operating systems such
as Windows.  Disabled is a better default.
Enable DPP2, support was added in 2.10 and there is no downside

Signed-off-by: Rick Farina <zerochaos@gentoo.org>