python{,-any}-r1.eclass: use python_has_version in examples
The python_check_deps() examples given in the documentation of several
functions and variables in python-r1.eclass and python-any-r1.eclass are
incorrect. The default behavior of has_version() is to check the
*runtime* system (RDEPEND), but python_check_deps() is meant to check
that the *build* system (BDEPEND) has a suitable Python environment. The
examples fail to do this and thus likely are leading developers to write
incorrect python_check_deps() functions in their ebuilds. Thankfully,
python-utils-r1.eclass supplies a python_has_version() function that
defaults to checking the build system and implements some other
enhancements that make it nicer for use in python_check_deps(). Change
the examples to use python_has_version.

Closes: https://bugs.gentoo.org/835462
Signed-off-by: Matt Whitlock <gentoo@mattwhitlock.name>
Signed-off-by: Michał Górny <mgorny@gentoo.org>