python*-r1.eclass: introduce python_get_any_deps() function
While very powerful, the existing python_check_deps() mechanism is
clunky to use, as it requires duplicating the dependency block passed to
python_gen_any_dep() but using shell syntax that is ugly and may admit
subtle differences in the logic. Indeed, the examples in the eclassdoc
were wrong: all of them were calling has_version without -b, so they
were checking the wrong sysroot, and one of them was inadventently
missing a USE-conditional that was present in the corresponding BDEPEND.

As an optional alternative to python_check_deps(), introduce an ebuild-
defined python_get_any_deps() function that prints to stdout the same
dependency specification as the ebuild passes to python_gen_any_dep().
This is easier to use correctly and should cover practically all cases.
Update the examples in the eclassdoc to use this new alternative instead
of python_check_deps().

Closes: https://bugs.gentoo.org/835462
Signed-off-by: Matt Whitlock <gentoo@mattwhitlock.name>