sys-process/cronie: fix cronie pidfile path
cronie's build system is using runstatedir to determine
where to place the PID file. On one of my systems,
this path _happens_ to exist because of thermald,
but it won't usually exist.

We should set runstatedir to be "${EPREFIX}"/run
to store the PID file in the place the init script expects
(cronie hardcodes the value of runstatedir into the binary
to find where its pidfile is).

Closes: https://bugs.gentoo.org/835814
Bug: https://bugs.gentoo.org/685306
Signed-off-by: Sam James <sam@gentoo.org>