net-libs/glib-networking: Avoid gnutls[pkcs11] dep without USE=test
Commit 7068209eac9 ("net-libs/glib-networking: gnutls needs pkcs11
support") added a USE dependency on net-libs/gnutls[pkcs11] in response
to a build problem. The appropriate fix was to double check the sed in
src_prepare() whose purpose is clearly noted as avoiding the pkcs11
dependence. It turns out the `subdir` line it attempted to modify had
been moved. With that resolved, we can continue avoiding the pkcs11
dependence.

Also revbump as required when runtime dependencies change, as should
have been done in the aforementioned commit.

Revert "net-libs/glib-networking: gnutls needs pkcs11 support"

This reverts commit 7068209eac9441762c2ad3ba1fe64d392e22e68b.

Closes: https://bugs.gentoo.org/835833
Signed-off-by: Matt Turner <mattst88@gentoo.org>