games-fps/urbanterror: Bump to 4.3.4_p20211208
This bump filters "-fstrict-aliasing". If it's enabled the game doesn't
connect to remote engines anymore. An upstream bug is filed as
https://github.com/mickael9/ioq3/issues/40

It also adds the trailing slash to the HOMEPAGE and updates the upstream
inside metadata.xml: The original one (Barbatos/...) is redirection to
FrozenSand/... (which looks lgtm), but as we actually use another engine
here it makes way more sense to point to that repo instead.

Signed-off-by: Nils Freydank <holgersson@posteo.de>