sys-apps/openrc: use seedrng for seeding the random number generator
This backports upstream's 076c2552aeff ("Use seedrng for seeding the
random number generator"), because it's not clear that an OpenRC release
is coming any time soon, and this is a security issue. In addition,
Alpine has already backported this to their equivalent of ~, so we
follow suit here.

Closes: https://bugs.gentoo.org/835926
Package-Manager: Portage-3.0.30, Repoman-3.0.3
Signed-off-by: Jason A. Donenfeld <zx2c4@gentoo.org>