x11-drivers/nvidia-drivers: enable static-libs USE by default
Now that hwloc depends on this by "default" when VIDEO_CARDS=nvidia,
not being enabled by default tend to be an annoyance.

This only has a cheap build-only dependency on libX11/Xext, is tiny,
and doesn't take long to build. Only makes some sense to disable if
wayland-only or headless without X (albeit this is why the USE exists
and isn't unconditional, /could/ be merged with USE=X although that
USE has more implications.. may revisit later).

Bug: https://bugs.gentoo.org/836004
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>