sys-apps/sysvinit: add 3.02
Note this drops the 'shutdown' patch given QEMU has been fixed
(the impetus for the original bug) and the patch no longer applies --
to rebase it, we'd need to adapt the new translations and then always
have to regenerate them too. Not worth it given the original problem
is gone.

(Although, that said, it looks like the translations aren't
bundled generated anyway, so we ended up adding USE=nls for po4a for now.)

Bug: https://bugs.gentoo.org/449354
Signed-off-by: Sam James <sam@gentoo.org>