www-client/seamonkey: webrtc fixes
* Fixes webrtc build for PPC64 through a patch.
* Makes webrtc disabled by default through an USE flag.

Seamonkey's webrtc should work, but there are a few drawbacks
that one should consider against using it. A plugin is needed
to activate it, as Seamonkey lacks an interface to do this.
Seamonkey deactivates it by default because of severe privacy
issues that could arise when not used properly. Another drawback
is that many webrtc-enabled websites block Seamonkey through an
User-Agent method, which would need a way to work around.

Signed-off-by: Myckel Habets <gentoo-bugs@habets-dobben.nl>