profiles: unmask vtk-9.1
The package has been masked for almost 4 months now for testing.
I've been building opencascade and freecad with this version since
then, without any big issues coming up.
If there are still single package versions of packages which fail to
build with this version, consider adjusting the depstring and relying
on 9.0.3, which I don't plan to drop anytime soon.

Signed-off-by: Bernd Waibel <waebbl-gentoo@posteo.net>
Closes: https://github.com/gentoo/gentoo/pull/24869
Signed-off-by: Sam James <sam@gentoo.org>