sys-libs/libcap-ng: add 0.8.3
Keep the Fedora patch for now (one last time?) which reverts a behaviour
change until we're a bit more certain consumers are fixed. Given Fedora
is still shipping it (and the upstream maintainer == Fedora maintainer),
I'd rather wait a bit longer before dropping.

That said, gnome-keyring is fixed since 40, and it was the only
known breakage AFAIK.

Bug: https://bugs.gentoo.org/759550
Signed-off-by: Sam James <sam@gentoo.org>