sys-libs/glibc: filter -fcf-protection
Set USE=cet if you want this. glibc can't be built with this *everywhere*,
and the configure option (controlled by USE=cet) sets it for the components
for which it works.

It's just like SSP and PIE. You can't force it on all of glibc, and we have
mechanisms to do it properly (USE=cet).

Closes: https://bugs.gentoo.org/830454
Signed-off-by: Sam James <sam@gentoo.org>