linux-info.eclass: Call ebegin, properly close with eend
A recent QA check added to portage informs of eend being called without
a preceding call to ebegin.

This warning was emitted during pkg_setup for net-vpn/openvpn, and was
traced back to the check_extra_config() function in linux-info.eclass.

The preference is here to call ebegin (instead of dropping the lone
eend) and in each of the possible exit branches for check_extra_config()
call eend appropriately.

Reported-by: Sam James <sam@gentoo.org>
Signed-off-by: Thomas Bracht Laumann Jespersen <t@laumann.xyz>