app-i18n/ibus: Replace IUSE=+gtk with IUSE=+gui
Now that we have separated IUSE=gtk3 from IUSE=gtk, IUSE=gtk only
enables the UI, so we rename it to IUSE=gui to be more in line with
Gentoo policies.

Additionally, move the gtk dependencies out of the gui? ( ... ) block,
since the functionality they control (namely the IM modules) can be
enabled independently of the UI.

Signed-off-by: Matt Turner <mattst88@gentoo.org>