gnome-base/gnome-shell: Require gtk3 and gtk4 support in ibus
The IUSE=gtk flag on ibus was split into IUSE=gui and IUSE=gtk3, so we need to
update this. I'm not confident in what gnome-shell actually requires, but since
it depends on x11-libs/gtk+:3 and gui-libs/gtk:4 already and both are enabled
by default in ibus, I'm going to play it safe and require both.

Signed-off-by: Matt Turner <mattst88@gentoo.org>