media-video/wireplumber: backport two minor fixes
The first fix deals with handling of device removal and is included for
correctness sake, since it appears to be the right thing for WirePlumber
to do.

The second fix implements and documents D-Bus disabling for
WirePlumber which must be done when attempting to run WirePlumber as
a system service. This does not mean it's a recommended or even fully
supported configuration but there's people who want it. And having it
available should not cause any harm to other users.

Signed-off-by: Niklāvs Koļesņikovs <89q1r14hd@relay.firefox.com>