profiles: unmask split pulseaudio packages
This is a move towards making it easier for users to choose
between pulseaudio and pipewire for their audio needs.

We now have:
- media-sound/pulseaudio (a transitory metapackage, installs no files, just
enforces deps with USE flags);
- media-libs/libpulse (the pulseaudio libraries, needed by clients using
pulseaudio's protocol to speak to pipewire, but is not the pulseaudio daemon)
- media-sound/pulseaudio-daemon (the pulseaudio daemon, to be used instead
of pipewire if desired)

While these are technically RC/pre-releases (and I have a strong reluctance
against keywording those usually), we have a good relationship
with upstream who report they've been quiet developmentwise since these tags,
these versions have been tested & work well, and this will help to improve
the experience for Gentoo users in choosing which audio server they want. So,
all in all, a win.

Bug: https://bugs.gentoo.org/536780
Bug: https://bugs.gentoo.org/744622
Bug: https://bugs.gentoo.org/827546
Bug: https://bugs.gentoo.org/841494
Signed-off-by: Sam James <sam@gentoo.org>