autotools.eclass: drop ROOT=/ in has_version call
Noticed when doing a read-through of the eclass (which is how I noticed
c8e74a7dfe477dea008548553141f083c5d03782 too).

Forcing ROOT=/ is, at best, going to confuse matters, and at worst,
do entirely the wrong thing. In EAPI 5 and 6, we had --host-root (which
we use in the eclass), and in EAPI 7+, we have -b/-d (which we use in
the eclass too). The ROOT= setting was there for pre-EAPI 5 times.

The aforementioned toggles (--host-root and then later on, -b/-d) are the
correct methods to specify (B)ROOT for has_version. Using one of those
with ROOT= have undefined behaviour, so drop it.

Bug: https://bugs.gentoo.org/312687
Thanks-to: James Le Cuirot <chewi@gentoo.org>
Thanks-to: Ulrich Müller <ulm@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>