dev-java/auto-value: new package, add 1.11.0
Preliminary version with truth and compile-testing bundled.

Annotation processor for @AutoValue annotations. It is needed for
building packages like dev-java/testparapeterinjector, dev-java/truth or
dev-java/compile-testing.

An example how to use it can also be seen in src_test() of this ebuild.
Further information in 'man 1 javac' (look for '-processorpath') and in
the manual.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>