dev-lang/ruby: run the full test suite
In previous ruby versions the full test suite could only be run against
an installed ruby version, but this is no longer the case, and in fact
the tests that we installed with the "rubytests" USE flag no longer
worked as-is. Instead, "make check" can be used to run the full suite
against the currently built ruby version.

In order for these checks to pass we also need to explicitly avoid
building the win32 extensions.

Thanks to chewi for investigating and suggesting improvements.

Signed-off-by: Hans de Graaff <graaff@gentoo.org>