media-sound/xmms2: add 0.9.1
Special thanks to slyfox for handling a few things upstream
and making a release after so long.

Mostly same as previous snapshot with patches but with several
of these upstreamed and now have a proper release tarball
including s4 submodule and updated waf to simplify the ebuild.

Also refreshed older patches (none are new).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>