media-libs/libepoxy: keyword 1.5.9-r1 for ~loong
Tests passed on real hardware with minor tweak; test/dlwrap.c needs to
be taught that "GLIBC_2.36" is also a possible version for dlsym. As the
loong glibc port is yet to be merged, the version could change, so not
submitting it at the moment; glx_alias_prefer_same_name and glx_gles2
are expected to fail.

Signed-off-by: WANG Xuerui <xen0n@gentoo.org>