eclass/libtool.eclass: Allow forcing the slibtool wrapper
With most builds the r{d,c,dc}libtool symlinks will result in the
correct behavior. These symlinks will use the builds own local copy
of the GNU libtool script created by the configure process to see
if the build enabled shared, static or both. However some builds do
not have a standard configure process and lack this script where these
slibtool varieties will then fail with flow errors.

The solution is to force the correct slibtool variety that the build
expects and this adds a new elibtool() function and LIBTOOL eclass
variable to accomplish that. Users should set their desired libtool
implementation via the LIBTOOL, MAKE and MAKEFLAGS variable as
described in the gentoo wiki and the eslibtool() function with modify
these variables if set accordingly by using the -shared, -static or no
argument.

In most cases where the build systems uses the standard configure
process the eslibtool() should not be used and in the case the LIBTOOL
variable is unset or a blank variable it will be a no opt.

Signed-off-by: orbea <orbea@riseup.net>