dev-util/dialog: Use eslibtool()
When using rlibtool the build will fail because there is no libtool
script for slibtool to determine if the build is shared or static.

Forcing one of the hardcoded slibtool wrappers will allow the build
to succeed.

When using GNU libtool and the LIBTOOL environment variable is unset
there will be no change in behavior.

Bug: https://bugs.gentoo.org/778047
Signed-off-by: orbea <orbea@riseup.net>