media-video/wireplumber: bump to 0.4.10
In addition to bumping to the just released 0.4.10 version, this commit
adds BDEPEND on sys-devel/gettext because turns out that since libintl
was added, Meson build system would end up calling xgettext to process
the translation files during build.

Apart from the previously missing gettext dependency, this is a trivial
copy of the wireplumber-9999 ebuild. Should be equally trivial bump.

Signed-off-by: Niklāvs Koļesņikovs <89q1r14hd@relay.firefox.com>
Closes: https://github.com/gentoo/gentoo/pull/25421
Signed-off-by: Sam James <sam@gentoo.org>