sys-auth/polkit: add 0.120_p20220509 (unkeyworded)
Unkeyworded for now until tomorrow probably. (Nothing major changed
but it's late/early and I don't want to take a risk and then go to bed.)

(In terms of the snapshot, the only notable things are that upstream
dropped autotools, but we were using meson for a little while now anyway,
and a bunch of the patches we were including to a previous snapshot
are now rolled into this one.)

I'd not bothered looking much into the test situation given the bug
flagged in the ebuild (the dist tarballs don't contain the test
dependencies needed and they don't use meson subprojects so we
can't provide them ourselves either)

... but a bug (with a dodgy premise -- because you shouldn't be
setting USE=test manually anyway; RESTRICT="test" is currently in ebuild anyhow)
exposed the fact that if you _do_ try to run tests right now, they end up
failing because of our musl patch which still includes config.h even though
we're doing a meson build.

So, a double win:
1. Tests are now wired up and run for snapshots at least! Yay!
2. Our musl patch is now cleaner and ready for another attempt
at upstream submission.

Closes: https://bugs.gentoo.org/843647
Signed-off-by: Sam James <sam@gentoo.org>