dev-util/mingw64-runtime: unset toolchain vars when cross-compiling
i.e. it could be CC=x86_64-pc-linux-gnu-gcc / clang which is
what's wanted for building tools, but after we need to discard
that to switch to x86_64-w64-mingw32-gcc (unless CHOST is already
mingw as it would be correct).

Ideally would need a unified way to do this with override variables
for users to specify the alternate toolchain variables.

The -Wl,--hash-style=* filter is likely not necessary given
strip-unsupported-flags already strip it, but if(?) CC was
set it may possibly have messed with that -- albeit will keep
it as a safety/informational.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>