dev-util/clippy: Fix clippy install with slibtool
Installing a file created by libtool is not portable for slibtool and
this will result in installing the slibtool shell wrapper script for
clippy and not the clippy binary.

This additionally fails when compiling net-misc/frr which will use the
shell wrapper script that does not work outside of the build directory.

The 'make install' target is not appropriate for --enable-clippy-only
and fixing that seems like significant work so just use shell instead to
find the correct file.

I bumped the ebuild revision because this causes a build failure in frr.

Bug: https://bugs.gentoo.org/830087
Signed-off-by: orbea <orbea@riseup.net>