sci-libs/gdal: work around jpeg/png bugs in 3.5.0 for now
Force on jpeg & png for now. It's not ideal but reported both upstream
and I lack the gdal internals knowledge to deduce whether the various
plugins which depend on each of them are functional (enough?) without
each of them.

They're both pretty small dependencies and hard to avoid in general,
although I accept it's an inconvenience, some of this stuff was fairly
expected & unavoidable with the big port to CMake.

Bug: https://github.com/OSGeo/gdal/issues/5741
Bug: https://github.com/OSGeo/gdal/issues/5742
Bug: https://bugs.gentoo.org/844751
Signed-off-by: Sam James <sam@gentoo.org>