gnome-base/gnome-shell: revbump for Polkit patch
As Matt points out, we might end up installing gnome-shell w/ an older
polkit, then upgrading polkit, and then nobody provides the autoref
bits.

Given it can go both ways, depend on a newer Polkit which always
has the autoref bits, and add the patch to said new version.

(Also, revbump -r1 -> r2 with no patch, for consistency, as people
will have likely merged the -r1 w/ the polkit build fix, but in theory
could (but unlikely!) downgrade polkit at some point, and it's
just a lot easier to handle bugs this way if we know everything is
in a sane state.)

Bit annoyed I didn't think it through fully as I went back and forth
a few times and decided on not, wrongly.

Bug: https://bugs.gentoo.org/844919
Signed-off-by: Sam James <sam@gentoo.org>