media-gfx/blender: Remove option to build standalone cycles binary
After talking to upstream, it seems like this option is not meant to be
use to ship any binaries. It is simply there as an example cycles
project that developers can easily modify if they want to quickly code
in new features and test them in cycles (without having to go through
blender).

Closes: https://bugs.gentoo.org/828880
Signed-off-by: Sebastian Parborg <darkdefende@gmail.com>