media-video/wireplumber: add 0.4.10-r2 and copy loong to 9999
This commit backports a fix for device reservation protocol where
multiple devices may get clashing name because they all get the same
name assigned to them while being reserved. It appears the fix may
need another commit to function, so adding the NULL fix pre-emptively.
And because that one introduced a new bug, adding a follow up fix
backport for that.

Unfortunately the current upstream fix for this breaks the fix for a
small audio glitch that may happen when changing the default device,
which as applied on Gentoo in the 0.4.10-r1 ebuild. This in effect
means no longer applying the corresponding patch. But the benefit
of not having potential device name clashes should outweigh that.

Finally 0.4.10-r1 was keyworded for ~loong but that's liable to getting
lost on version bumps, so copying it to the 9999 in an inactive state.

Signed-off-by: Niklāvs Koļesņikovs <89q1r14hd@relay.firefox.com>
Closes: https://github.com/gentoo/gentoo/pull/25574
Signed-off-by: Sam James <sam@gentoo.org>