games-emulation/pcsx2: update live
Upstream added a generic function to ease unbundling which
has USE_SYSTEM_LIBS to enable all at once, but not everything
is covered yet (e.g. no cubeb/chdr/glslang).

Also switch to using aliases to have less files to modify.

I expect this to break again soon'ish as upstream continues
to work on this (please report or ping on IRC, I do not try
to build this often to check).

Closes: https://bugs.gentoo.org/847343
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>