media-video/wireplumber: fix a bad mistake in 0.4.10-r2 by doing -r3
Somehow git format-patch ended up producing a revert for an unrelated
commit which either randomly or perhaps via some copy and paste error
happened to be a revert of the only patch that 0.4.10-r1 was applying.

Because of that wireplumber-0.4.10-r2 incorrectly dropped it. This is
now fixed in 0.4.10-r3 by reinstating the removed patch as well as
adding in the one that was meant to be present in -r2 but had been
incorrectly generated due to probably wrong git format-patch command.

Signed-off-by: Niklāvs Koļesņikovs <89q1r14hd@relay.firefox.com>