sci-libs/hdf5: fix cmake include dir regression
An attempt to build hdf5-1.12.2-r1 with USE='fortran' on a system where
an older version of the package has already been installed will fail due
to system headers taking precedence over the ones present in the source
directory.

This exact issue (#808633) has already been fixed in 225b7ee for 1.12.1,
though the workaround somehow hadn't made it into the new ebuild. This
commit brings it in.

Bug: HDFGroup/hdf5#1027
Closes: https://bugs.gentoo.org/808633
Signed-off-by: Roman Beranek <roman.beranek@prusa3d.com>