dev-db/mariadb: apply upstream 10.6.x OpenSSL 3 patches (to 10.6.8-r1)
These are already queued on the upstream 10.6.x branch. Having a version
of MariaDB that's OpenSSL 3 capable is important given it unblocks
other packages and more wide testing.

(Also, while this isn't what we're doing, Fedora has been using
a bunch of OpenSSL 3 patches on top of 10.5.x for a while.)

Closes: https://bugs.gentoo.org/849593
Signed-off-by: Sam James <sam@gentoo.org>