dev-lang/python: drop long-obsolete SSP / -O3 workaround
Everyone in Gentoo (unless they unforce defaults) had, for years now,
-fstack-protector-all on, or nowadays, -fstack-protector-strong.

This means that it won't have been explicitly in *FLAGS and hence
the replacement won't have been triggered anyway.

Plus, the hardened conditional is irrelevant given the above
(it uses no more SSP than the default, because the default
is what hardened used to be, many years ago).

Bug: https://bugs.gentoo.org/50309
Signed-off-by: Sam James <sam@gentoo.org>