distutils-r1.eclass: Introduce DISTUTILS_USE_PEP517=no mode
Introduce a new DISTUTILS_USE_PEP517 value "no" that stands for
"no build system".  This is primarily meant to replace the legacy
distutils-r1 logic used for bootstrapping baseline PEP 517 packages.
At the same time, it provides a convenient replacement for some
of the uses of python-r1.

In this mode, the eclass does not add PEP 517-specific dependencies
and the default python_compile() is a no-op.  However, it does set
dependencies, REQUIRED_USE and enables sub-phase usage (with respect to
DISTUTILS_OPTIONAL).  It also permits using
distutils_enable_{sphinx,tests}.

Signed-off-by: Michał Górny <mgorny@gentoo.org>