sci-libs/hdf5: Set H5_DEFAULT_PLUGINDIR.
HDF5 sets is default plugin dir to /usr/local/hdf5/lib/plugin. It is
reasonable in most cases, but not very well for Gentoo.

In Gentoo, the lib directory is not always `lib`, and the prefix is not
always `/`. Therefore, it is more reasonable to set the default dir under
${EPREFIX}, and change `lib` to the right lib dir name.

Signed-off-by: Yuyi Wang <Strawberry_Str@hotmail.com>
Closes: https://github.com/gentoo/gentoo/pull/25814
Signed-off-by: Sam James <sam@gentoo.org>