dev-embedded/avr-libc: Bump to 2.1.0, drop IUSE="doc"
Bump to the new version and EAPI 8. There is an issue with upstream's
distfile signature for the HTML documentation, so dropping for now. IMHO we
should restore the flag and doc installation as soon as upstream's signature
is fixed.
Man pages are not affected as they are part of the main distfiles archive.

Bug: https://bugs.gentoo.org/837146
Signed-off-by: Nils Freydank <holgersson@posteo.de>
Signed-off-by: Ulrich Müller <ulm@gentoo.org>