app-admin/github-backup-utils: call python-any-r1_pkg_setup conditionally
As per example in the Python guide. Python is only needed for tests here,
but as-is, it'll check for Python on all emerges, which is a problem if
e.g. upgrading an old system, Python has partly been upgraded, but not
completely yet, then the emerge will die and the system may be left
in a bad state.

(i.e. the pkg_setup call should match the dependency.)

Signed-off-by: Sam James <sam@gentoo.org>