x11-drivers/nvidia-drivers: pin to egl-wayland-1.1.9 for 0/510
Upcoming egl-wayland-1.1.10 is known incompatible, meant to bump
it early to fix wayland vulkan issues with 0/515 but it requires
yet unreleased nvidia-drivers that'd have the missing symbol in
the vulkan-producer lib.

And so .10 will likely never work with old 0/510, and there should
be no reason to use 1.1.7 with it (barely tested combination).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>