sys-kernel/linux-headers: add 5.19_rc1
While we can certainly wait for the official 5.19 release, adding this
facilitates early loong testing across a wide variety of packages,
because 5.19 is the earliest Linux version with upstreamed loong
support. IUSE=experimental-loong is dropped while at it.

This package will be replaced with the proper 5.19 release once that one
is officially out.

This is ~loong-only, because there isn't significant changes for other
arches.

Reviewed-by: Sam James <sam@gentoo.org>
Signed-off-by: WANG Xuerui <xen0n@gentoo.org>