llvm.org.eclass: Control manpage type via eclass
Move switching between pregenerated and built manpages from ebuilds
to the eclass.  This is going to make the maintenance easier — rather
than having to switch to `build` on every version bump, then back
to `pregenerated` we will just bump the current version in eclass
after making the new manpage tarball.

Signed-off-by: Michał Górny <mgorny@gentoo.org>