linux-mod.eclass: Support module compression
The Linux kernel supports the compression of modules utilizing GZIP, XZ
and ZSTD.  Add support into linux-mod.eclass to support this for out of
tree modules utilizing the compression binary specified in the kernel
config.

Note that if the binary which provides the compression is not present on
the system the kernel would have failed to build with an error
indicating the missing binaries name.

Closes: https://bugs.gentoo.org/850130

Signed-off-by: Mike Pagano <mpagano@gentoo.org>