toolchain.eclass: BDEPEND on D for bootstrapping D
To build GCC 12+ w/ D, we need a D compiler. For now, depend
on GCC with D enabled. And give a hint to the PM that an older
GCC is likely to do the job to break the cycle, as those don't
need D to build their D support. Just in newer versions.

Closes: https://bugs.gentoo.org/840182
Signed-off-by: Sam James <sam@gentoo.org>