app-arch/zstd: add split-usr support
Done with some reluctance, as I don't really want to be introducing more of
these, but it blocks being able to turn on zstd by default safely in kmod,
and it's a simple enough fix that I can't justify breaking things - it'd
be gratuitious.

Not tested on a split-usr system though, so it's really best-effort.

(split-usr being "/usr is a separate partition and therefore libraries
need to be in /$(get_libdir) for boot to bring it up", and not anything
to do with "merged-usr").

Closes: https://bugs.gentoo.org/771078
Signed-off-by: Sam James <sam@gentoo.org>