net-firewall/nftables: add emergency pkg_preinst sanity check
The idea here is that we check just before merging whether the freshly
built `nft` binary can safely (pretend) reload the system ruleset.

A significant number of recent regressions have manifested in immediate
segfaults when doing this, so it's worth doing the safety check (it's
not as if it's niche or unlikely for a failure to occur at this point).

Those who want a failed check to be *fatal* can set NFTABLES_ABORT_ON_RELOAD_FAILURE=1.

Otherwise, it's just a warning.

Bug: https://bugs.gentoo.org/852662
Thanks-to: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>