x11-drivers/nvidia-drivers: add 515.49.05:0/vulkan
Note that even though it is newer (for now), it is not recommended
to use the 0/vulkan branch unless you are involved in vulkan
application development, upstream bugs may not get fixed (aka it
adds vulkan features on top of old drivers and rebase occasionally).

Please report if packaging issues, as don't check masked 0/vulkan much.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>